split minimap_rays and added free_2d_array
