+++ src/moving_bonus.c
@@ -6,19 +6,12 @@
 /*   By: sniemela <sniemela@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/04 16:44:46 by sniemela          #+#    #+#             */
-/*   Updated: 2025/04/07 15:29:29 by sniemela         ###   ########.fr       */
+/*   Updated: 2025/04/08 12:04:01 by sniemela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/cub3d.h"
 
-int		get_index_of_rov_and_col(t_data *data, int x, int y, enum dir_type gear)
-{
-	float	i;
-	(void)gear;
-	i = y * data->map_info.width + x;
-	return (i);
-}
 
 void	move_up_down(t_data *data)
 {
@@ -30,10 +23,10 @@ void	move_up_down(t_data *data)
 	{
 		x = data->player.x + data->player.d_x;
 		y = data->player.y + data->player.d_y;
-		i = get_index_of_rov_and_col(data, (int)(data->player.x), (int)(y), 0);
+		i = get_index_of_rov_and_col(data, (int)(data->player.x), (int)(y));
 		if (data->map[i] == FLOOR)
 			data->player.y = y;
-		i = get_index_of_rov_and_col(data, (int)(x), (int)(data->player.y), 0);
+		i = get_index_of_rov_and_col(data, (int)(x), (int)(data->player.y));
 		if (data->map[i] == FLOOR)
 			data->player.x = x;
 	}
@@ -41,54 +34,13 @@ void	move_up_down(t_data *data)
 	{
 		x = data->player.x - data->player.d_x;
 		y = data->player.y - data->player.d_y;
-		i = get_index_of_rov_and_col(data, (int)(data->player.x), (int)(y), 1);
+		i = get_index_of_rov_and_col(data, (int)(data->player.x), (int)(y));
 		if (data->map[i] == FLOOR)
 			data->player.y = y;
-		i = get_index_of_rov_and_col(data, (int)(x), (int)(data->player.y), 1);
+		i = get_index_of_rov_and_col(data, (int)(x), (int)(data->player.y));
 		if (data->map[i] == FLOOR)
 			data->player.x = x;
 	}
-}
-
-void	draw_minimap_rays(t_data *data)
-{
-	// float	ray_len;
-	int		i;
-	float	ray_angle;
-	float	ray_x;
-	float	ray_y;
-	float	map_x;
-	float	map_y;
-	int		xy;
-
-	i = 0;
-	mlx_delete_image(data->mlx, data->ray);
-	data->ray = mlx_new_image(data->mlx, WIDTH, HEIGHT);
-	ray_x = data->player.x;
-	ray_y = data->player.y;
-	ray_angle = data->player.angle - (33.0 / 180 * PI);
-	if (ray_angle >= 2 * PI)
-		ray_angle -= 2 * PI;
-	else if (ray_angle < 0)
-		ray_angle += 2 * PI;
-	while (i <= 66)
-	{
-		xy = get_index_of_rov_and_col(data, (int)(data->player.x), (int)(data->player.y), 1);
-		map_y = ray_y;
-		map_x = ray_x;
-		while (data->map[xy] == FLOOR)
-		{
-			map_x += cos(ray_angle) * 0.01;
-			map_y += sin(ray_angle) * 0.01;
-			xy = get_index_of_rov_and_col(data, (int)map_x, (int)map_y, 1);
-			mlx_put_pixel(data->ray, map_x * data->tile_mini - 5, map_y * data->tile_mini - 5, RED);
-		}
-		ray_angle += (1.0/180 * PI);
-		if (ray_angle >= 2 * PI)
-			ray_angle -= 2 * PI;
-		i++;
-	}
-	mlx_image_to_window(data->mlx, data->ray, data->player.x, data->player.y);
 }
 
 void	move_left_right(t_data *data)
@@ -101,10 +53,10 @@ void	move_left_right(t_data *data)
 	{
 		x = data->player.x - cos(data->player.angle + PI / 2) * MOVE_SPEED;
 		y = data->player.y - sin(data->player.angle + PI / 2) * MOVE_SPEED;
-		i = get_index_of_rov_and_col(data, (int)(data->player.x), (int)(y), 2);
+		i = get_index_of_rov_and_col(data, (int)(data->player.x), (int)(y));
 		if (data->map[i] == FLOOR)
 			data->player.y = y;
-		i = get_index_of_rov_and_col(data, (int)(x), (int)(data->player.y), 2);
+		i = get_index_of_rov_and_col(data, (int)(x), (int)(data->player.y));
 		if (data->map[i] == FLOOR)
 			data->player.x = x;
 	}
@@ -112,10 +64,10 @@ void	move_left_right(t_data *data)
 	{
 		x = data->player.x + cos(data->player.angle + PI / 2) * MOVE_SPEED;
 		y = data->player.y + sin(data->player.angle + PI / 2) * MOVE_SPEED;
-		i = get_index_of_rov_and_col(data, (int)(data->player.x), (int)(y), 3);
+		i = get_index_of_rov_and_col(data, (int)(data->player.x), (int)(y));
 		if (data->map[i] == FLOOR)
 			data->player.y = y;
-		i = get_index_of_rov_and_col(data, (int)(x), (int)(data->player.y), 3);
+		i = get_index_of_rov_and_col(data, (int)(x), (int)(data->player.y));
 		if (data->map[i] == FLOOR)
 			data->player.x = x;
 	}
@@ -156,13 +108,13 @@ void	movement(void *param)
 	data->view = raycaster(data);
 	mlx_image_to_window(data->mlx, data->view, 0, 0);
 	if (data->minimap)
-	move_player_image(data);
+		move_player_image(data);
 	if (data->minimap)
 		draw_minimap_rays(data);
 	mlx_set_instance_depth(data->view->instances, 3);
 	if (data->minimap)
 	{
-		 mlx_set_instance_depth(data->minimap->instances, 4);
+		mlx_set_instance_depth(data->minimap->instances, 4);
 		mlx_set_instance_depth(data->mini_p_img->instances, 5);
 		mlx_set_instance_depth(data->ray->instances, 6);
 	}
